+++ LinearEquations/LinearEquations/ISeidelsMethos.cs
@@ -3,7 +3,7 @@ namespace LinearEquations
 {
     public interface ISeidelsMethos
     {
-        int[] calculate(int[,] array, int size);
+        double[] calculate(double[,] array, int size);
         // Cтатический метод calculate принимает на вход
         // матрицу коэффициентов размера size*size+1, 
         // само значение размера матрицы size.
